importing spandsp 0.0.4pre18 (same as snapshot 20080305)

+++ src/gsm0610_local.h
@@ -25,7 +25,7 @@
  * This code is based on the widely used GSM 06.10 code available from
  * http://kbs.cs.tu-berlin.de/~jutta/toast.html
  *
- * $Id: gsm0610_local.h,v 1.7 2007/01/03 14:15:35 steveu Exp $
+ * $Id: gsm0610_local.h,v 1.8 2008/03/03 15:29:40 steveu Exp $
  */
 
 #if !defined(_GSM0610_LOCAL_H_)
@@ -140,15 +140,15 @@ static __inline__ int16_t gsm_abs(int16_t a)
 static __inline__ int16_t gsm_asr(int16_t a, int n)
 {
     if (n >= 16)
-        return  -(a < 0);
+        return  (int16_t) (-(a < 0));
     /*endif*/
     if (n <= -16)
         return  0;
     /*endif*/
     if (n < 0)
-        return a << -n;
+        return (int16_t) (a << -n);
     /*endif*/
-    return  a >> n;
+    return  (int16_t) (a >> n);
 }
 /*- End of function --------------------------------------------------------*/
 
@@ -158,12 +158,12 @@ static __inline__ int16_t gsm_asl(int16_t a, int n)
         return  0;
     /*endif*/
     if (n <= -16)
-        return  -(a < 0);
+        return  (int16_t) (-(a < 0));
     /*endif*/
     if (n < 0)
         return  gsm_asr(a, -n);
     /*endif*/
-    return  a << n;
+    return  (int16_t) (a << n);
 }
 /*- End of function --------------------------------------------------------*/
